Splitting commands in meaninful folders
