+++ src/main/java/com/dis/bot/commands/character/CharacterCreateCommand.java
@@ -1,5 +1,6 @@
-package com.dis.bot.commands;
+package com.dis.bot.commands.character;
 
+import com.dis.bot.commands.SlashCommand;
 import com.dis.bot.repository.Characters;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
 import discord4j.core.object.command.ApplicationCommandInteractionOption;
@@ -8,36 +9,37 @@ import org.springframework.stereotype.Component;
 import reactor.core.publisher.Mono;
 
 @Component
-public class DamageCommand implements SlashCommand {
+public class CharacterCreateCommand implements SlashCommand {
 
     private final Characters characters;
 
-    public DamageCommand(Characters characters){
+    public CharacterCreateCommand(Characters characters){
         this.characters = characters;
     }
 
     @Override
     public String getName() {
-        return "damage";
+        return "character-create";
     }
 
     @Override
     public Mono<Void> handle(ChatInputInteractionEvent event) {
-        Long dmg = event.getOption("dmg")
-            .flatMap(ApplicationCommandInteractionOption::getValue)
-            .map(ApplicationCommandInteractionOptionValue::asLong)
-            .get();
+        String memberName = event.getInteraction().getMember().get().getUsername();
 
         String characterName = event.getOption("name")
+            .flatMap(ApplicationCommandInteractionOption::getValue)
+            .map(ApplicationCommandInteractionOptionValue::asString)
+            .get();
+
+        Long hp = event.getOption("hp")
                 .flatMap(ApplicationCommandInteractionOption::getValue)
-                .map(ApplicationCommandInteractionOptionValue::asString)
+                .map(ApplicationCommandInteractionOptionValue::asLong)
                 .get();
 
-
-        var character = characters.applyDamage(characterName, dmg);
+        var character = characters.getOrCreate(characterName, hp, memberName);
 
         return  event.reply()
             .withEphemeral(false)
-            .withContent(String.format("%s current HP now is %d ", character.getName(), character.getCurrentHealthPoints()));
+            .withContent(String.format("Character, %s was created with %d HP", character.getName(), character.getHealthPoints()));
     }
 }
