+++ src/main/java/com/dis/bot/commands/hp/ShowHpCommand.java
@@ -1,5 +1,6 @@
-package com.dis.bot.commands;
+package com.dis.bot.commands.hp;
 
+import com.dis.bot.commands.SlashCommand;
 import com.dis.bot.repository.Characters;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
 import org.springframework.stereotype.Component;
