+++ src/main/java/com/dis/bot/commands/init/InitiativeCommand.java
@@ -1,5 +1,6 @@
-package com.dis.bot.commands;
+package com.dis.bot.commands.init;
 
+import com.dis.bot.commands.SlashCommand;
 import com.dis.bot.repository.Characters;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
 import discord4j.core.object.command.ApplicationCommandInteractionOption;
@@ -8,31 +9,36 @@ import org.springframework.stereotype.Component;
 import reactor.core.publisher.Mono;
 
 @Component
-public class IniCommand implements SlashCommand {
+public class InitiativeCommand implements SlashCommand {
 
     private final Characters characters;
 
-    public IniCommand(Characters characters){
+    public InitiativeCommand(Characters characters){
         this.characters = characters;
     }
 
     @Override
     public String getName() {
-        return "ini";
+        return "initiative";
     }
 
     @Override
     public Mono<Void> handle(ChatInputInteractionEvent event) {
-        Long bonus = event.getOption("initiative-bonus")
+        Long initiative = event.getOption("initiative")
+            .flatMap(ApplicationCommandInteractionOption::getValue)
+            .map(ApplicationCommandInteractionOptionValue::asLong)
+            .get();
+
+        String characterName = event.getOption("name")
                 .flatMap(ApplicationCommandInteractionOption::getValue)
-                .map(ApplicationCommandInteractionOptionValue::asLong)
+                .map(ApplicationCommandInteractionOptionValue::asString)
                 .get();
 
-        var memberName = event.getInteraction().getMember().get().getUsername();
-        var character = characters.rollD20InitiativeFromMemberWithBonus(memberName, bonus);
+
+        var character = characters.setInitiative(characterName, initiative);
 
         return  event.reply()
             .withEphemeral(false)
-            .withContent(String.format("%s initiative total as %d ", character.getName(), character.getInitiative()));
+            .withContent(String.format("%s initiative set as %d ", character.getName(), character.getInitiative()));
     }
 }
