+++ src/main/java/com/dis/bot/commands/init/QuickInitCommand.java
@@ -1,5 +1,6 @@
-package com.dis.bot.commands;
+package com.dis.bot.commands.init;
 
+import com.dis.bot.commands.SlashCommand;
 import com.dis.bot.repository.Characters;
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent;
 import discord4j.core.object.command.ApplicationCommandInteractionOption;
@@ -8,17 +9,17 @@ import org.springframework.stereotype.Component;
 import reactor.core.publisher.Mono;
 
 @Component
-public class InitiativeCommand implements SlashCommand {
+public class QuickInitCommand implements SlashCommand {
 
     private final Characters characters;
 
-    public InitiativeCommand(Characters characters){
+    public QuickInitCommand(Characters characters){
         this.characters = characters;
     }
 
     @Override
     public String getName() {
-        return "initiative";
+        return "init";
     }
 
     @Override
@@ -28,13 +29,8 @@ public class InitiativeCommand implements SlashCommand {
             .map(ApplicationCommandInteractionOptionValue::asLong)
             .get();
 
-        String characterName = event.getOption("name")
-                .flatMap(ApplicationCommandInteractionOption::getValue)
-                .map(ApplicationCommandInteractionOptionValue::asString)
-                .get();
-
-
-        var character = characters.setInitiative(characterName, initiative);
+        var memberName = event.getInteraction().getMember().get().getUsername();
+        var character = characters.setInitiativeFromMember(memberName, initiative);
 
         return  event.reply()
             .withEphemeral(false)
